Backend/requirements.txt
langchain==0.3.25
langchain_community==0.3.23
langchain_google_genai==2.1.4
langchain_groq==0.3.2
langgraph==0.4.3

langgraph-checkpoint-postgres==2.0.21

parsedatetime==2.6
apscheduler==3.11.0

email-validator==2.2.0

fastapi[standard]==0.115.12
uvicorn[standard]==0.34.2
gunicorn==23.0.0

sqlalchemy==2.0.41

python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==3.2.2

psycopg[binary,pool]==3.2.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
weaviate-client[agents]>=4.4.0
pydantic~=2.11.5
typing_extensions~=4.14.0
pytest~=8.4.0


# langsmith<0.3.0,>=0.2.0